docs: In Chinese and doesn't prompt an auto translate (#21428)
* In Chinese and doesn't prompt an auto translate

All the pages on the website are in English, whereas this page is in Chinese. Creating an abrupt experience. 
The work with us page is not prompting the auto-translate feature on chrome or Edge Beta, or a person unfamiliar with the language it might be annoying to figure out which one it is

* Apply suggestions from code review

The new email looks like a real id

Co-Authored-By: 偏右 <afc163@gmail.com>

Co-authored-by: 偏右 <afc163@gmail.com>